+++ ctwin32/bcrypt.py
@@ -40,7 +40,7 @@ from .wtypes import (
 from . import ref, fun_fact
 from .ntdll import raise_failed_status, STATUS_INVALID_SIGNATURE
 
-_bcr = ctypes.WinDLL("bcrypt.dll")
+_bcr = ctypes.WinDLL("bcrypt.dll", use_last_error=True)
 
 ################################################################################