+++ changelog.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- switched to using ctypes' last error shadow copy
+
 ### Fixed
 
 - fixed string length handling in sample dump_ver_res.py